requirements.txt
asgiref==3.2.5
BaseAPI==0.1.1
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.5
django-cors-headers==3.1.1
django-extensions==2.2.8
django-filter==2.2.0
django-import-export==1.2.0
django-otp==0.8.1
djangorestframework==3.10.3
et-xmlfile==1.0.1
idna==2.8
jdcal==1.4.1
jsonfield==3.1.0
Markdown==3.1.1
MarkupPy==1.14
numpy==1.18.1
odfpy==1.4.0
openpyxl==3.0.0
pandas==1.0.3
Pillow==7.1.2
pyfcm==1.4.7
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.1
tablib==0.14.0
urllib3==1.25.6
xlrd==1.2.0
xlwt==1.3.0
